+++ R/ioc.R
@@ -17,6 +17,10 @@
 #' specialists in the assessment of criterion-referenced test item validity.
 #' *Dutch Journal of Educational Research, 2*, 49-60.
 #'
+#' Turner, R. C., & Carlson, L. (2003). Indexes of item-objective congruence
+#' for multidimensional items. *International Journal of Testing, 3*(2), 163-171.
+#' \doi{10.1207/S15327574IJT0302_5}
+#'
 #' @examples
 #' df <- data.frame(
 #'   item = rep("I1", 6),